requirements.txt
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.0.1
click==8.1.3
colorama==0.4.6
crypto==1.4.1
cryptography==39.0.0
Deprecated==1.2.13
firebase==3.0.1
Flask==2.2.2
gcloud==0.18.3
googleapis-common-protos==1.58.0
gunicorn==20.1.0
httplib2==0.21.0
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
jwcrypto==1.4.2
MarkupSafe==2.1.2
Naked==0.1.32
oauth2client==4.1.3
protobuf==4.22.0rc2
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
pycryptodome==3.17
pycryptodomex==3.17
pyparsing==3.0.9
python-jwt==4.0.0
PyYAML==6.0
requests==2.28.2
requests-toolbelt==0.10.1
rsa==4.9
shellescape==3.8.1
six==1.16.0
sseclient==0.0.27
urllib3==1.26.14
Werkzeug==2.2.2
wrapt==1.14.1
python-dotenv==0.21.1